Lesson 20
Lesson about useLoader hook.
Used to pre load images and/or models.
In case of images, they can be used to be applied to materials as textures.
useLoader hook needs the three.js loader as first argument (i.e. TextureLoader) and the url of where the file is (i.e. ./images/grid.png)